Refactor review and trace boundaries
Move review diff parsing behind ReviewDiffParser with explicit contracts and docs, then convert low-level trace recording and console formatting to class-based boundaries. Update callers, package exports, and focused tests to avoid one-off wrapper functions.
