+++ src/core/agent/context/run-context-builder.ts
@@ -1,6 +1,6 @@
 import { resolve } from 'node:path';
 import { createToolRegistry } from '@/core/tools/registry.js';
-import { createTraceRecorder } from '@/core/trace/recorder.js';
+import { TraceRecorder } from '@/core/trace/index.js';
 import { createBudget } from '@/core/utils/budget.js';
 import { buildSystemPrompt } from '@/core/prompts/system-prompt.js';
 import { logger as defaultLogger } from '@/core/utils/logger.js';
@@ -18,7 +18,7 @@ import type { ChatMessage } from '@/core/llm/types.js';
 export class AgentRunContextBuilder {
   static create(options: BuildAgentRunContextArgs): AgentRunContext {
     const registry = createToolRegistry(options.tools);
-    const trace = createTraceRecorder();
+    const trace = new TraceRecorder();
     const now = () => new Date().toISOString();
     const maxSteps = options.maxSteps ?? DEFAULT_MAX_STEPS;
     const toolNames = registry.names();
