+++ src/core/review/index.ts
@@ -1,0 +1,7 @@
+export { ReviewDiffParser } from './service.js';
+export type {
+  ReviewDiffFile,
+  ReviewDiffHunk,
+  ReviewDiffLine,
+  ReviewFileStatus,
+} from './types.js';
