+++ src/core/trace/index.ts
@@ -1,0 +1,3 @@
+export { TraceConsoleFormatter } from './format.js';
+export { TraceRecorder } from './recorder.js';
+export type { TraceRecordSink } from './types.js';
