+++ src/core/trace/types.ts
@@ -1,0 +1,7 @@
+import type { TraceEvent } from '@/core/types.js';
+
+export type TraceRecordSink = {
+  record(event: TraceEvent): void;
+  getTrace(): TraceEvent[];
+  toJSON(): string;
+};
