+++ src/server/features/control-plane/types.ts
@@ -2,11 +2,11 @@ import type { AgentLoopEvent } from '@/core/runtime/loop/index.js';
 import type { DaemonOwnerRecord } from '@/core/runtime/daemon/index.js';
 import type { HeartbeatRunView, HeartbeatTaskView } from '@/core/heartbeat/index.js';
 import type { WorkspaceDescriptor } from '@/core/runtime/workspaces/index.js';
-import type { MemoryStatusView } from '../../../core/memory/types.js';
-import type { ReviewDiffFile } from '../../../core/review/diff-domain.js';
+import type { MemoryStatusView } from '@/core/memory/types.js';
+import type { ReviewDiffFile } from '@/core/review/index.js';
 import type { ProviderCredentialSource } from '@/core/runtime/credentials/index.js';
-import type { ReasoningEffort } from '../../../core/llm/types.js';
-import type { ChatSessionRetention } from '../../../core/chat/types.js';
+import type { ReasoningEffort } from '@/core/llm/types.js';
+import type { ChatSessionRetention } from '@/core/chat/types.js';
 
 export type ChatSessionView = {
   id: string;
